requirements.txt
database-connect== 0.1.66
matplotlib== 3.7.1
seaborn== 0.12.2
boto3==1.24.84
botocore-stubs==1.27.86
dill==0.3.5.1
dnspython==2.2.1
evidently==0.1.58.dev0
fastapi==0.78.0
from-root==1.0.2
httptools==0.5.0
imblearn==0.0
mypy-boto3-s3==1.24.76
pip-chill==1.0.1
pymongo==4.2.0
jinja2== 3.1.2
python-dotenv==0.21.0
PyYAML==5.4.1
types-s3transfer==0.6.0.post4
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
wincertstore==0.2
kneed== 0.8.2
xgboost==0.90
python-multipart==0.0.6
neuro_mf==0.0.5
numba==0.56.4
numpy==1.23.5

-e .
